+++ be/src/runtime/routine_load/data_consumer_group.h
@@ -128,7 +128,7 @@ private:
 
     size_t len_of_actual_data(const char* data);
 
-    size_t index_of_json_begin(const char* data);
+    long index_of_json_begin(const char* data);
 
 private:
     // blocking queue to receive msgs from all consumers
